moved config method to api controller
